Disable reset in PuzzlesPanel when progress is 0
To do this, I added a field to the TreeView store. The new field
specifies whether or not the icon should be shown. The
_reset_progress_file method also checks whether the icon exists before
prompting the user.

A lot more things in this commit:

_compute_progress_info was made a static method because it can be one.

The stock_id property of Pixbuf was actually deprecated so I switched to
the new icon_name property and created the GTK_ICON_VIEW_REFRESH
constant.
see (https://lazka.github.io/pgi-docs/Gtk-3.0/constants.html#Gtk.STOCK_REFRESH)

The i18n of the modal message did not use interpolation as it should.
This commit fixes that.