Fix GHA workflow so test failure is properly caught
The current GHA "Run the test suite" workflow will virtually
never fail, because run_tests.py never exits anything but 0. It
effectively runs the tests, but ignores the results. The tests
have actually been failing for months - likely since 983e138 -
but nobody noticed because of this problem.

This changes run_tests.py so it exits 1 if the tests fail, and
tweaks run-tests.yml so we will still kill xvfb before exiting
if the tests fail.

Signed-off-by: Adam Williamson <awilliam@redhat.com>