Adjust test usage of get_event_loop() for Python 3.14 changes
As of Python 3.14, get_event_loop() no longer automatically sets
up a new loop if there currently isn't one, instead it raises a
RuntimeError. This adjusts several usages of get_event_loop() in
the tests to work with the change. It also adds a loop setup step
to learn.py - currently it doesn't have one, but calls
asyncio.create_task, so it is implicitly relying on an earlier
test having set up an event loop. If you run learn.py alone, it
fails because of this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>